requirements.txt
# WS2812 LED control
rpi_ws281x>=1.3.1

# GPIO control for motor encoder and sensors
RPi.GPIO>=0.7.0

# I2C communication for motor HAT
smbus2>=0.4.1

# Additional dependencies
#numpy>=1.19.0  # numerical operations
